# [Perf] Use small max_num_batched_tokens for A100