This PR solves a performance regression issue on A100. Related PR: #17073 

|                 | Token throughput (Before this PR) | Token throughput (After this PR) |
|-----------------|-----------------------------------|----------------------------------|
| Llama 70B, TP=2 | 1054.01                           | 1438.49   (+36%)                       |
| Llama 70B, TP=4 | 3053.03                           | 2994.03 (-2%)                          |


vLLM launching command:
TP=2:
```
vllm serve meta-llama/Llama-3.1-70B-Instruct     --port 8000     -tp 2     --max-model-len 10000     --gpu-memory-utilization 0.95     --swap-space 0
```
TP=4:
```
vllm serve meta-llama/Llama-3.1-70B-Instruct     --port 8000     -tp 4     --max-model-len 10000     --gpu-memory-utilization 0.95     --swap-space 0
```

Benchmarking command:
```
python benchmark_serving.py --port 8000     --dataset-name sharegpt     --dataset-path ShareGPT_V3_unfiltered_cleaned_split.json     --num-prompts 400     --request-rate inf     --model meta-llama/Llama-3.1-70B-Instruct
```

<!--- pyml disable-next-line no-emphasis-as-heading -->