@hmellor 2025-05-09
It might make sense to move all these heuristics into the respective platform file?

As it is right now we have side effects where non-GPU systems (except TPU) are having their defaults set based on memory heuristics meant for GPUs. i.e. if you have a CPU platform with >70GB RAM you're get the large GPU defaults for `max_num_batched_tokens`...