@heheda12345 2025-05-09
Would `cudaDeviceProp.multiProcessorCount` be a better metric? We can get this metric from numba.