@MingzhenHan 2025-07-06
@KuntaiDu  Hello Kuntai, Why do we need to use smaller max_num_batched_tokens on A100? What is the reason for this? Looking forward to your guidance🙏